2do commit componentes
